+++ src/types/params.d.ts
@@ -1,4 +1,5 @@
 export type RVParams = {
 	min?: number;
 	max?: number;
+	description?: string;
 };